+++ DuoArtifacts/ExternalAPIs/IMoreDifficultiesApi.cs
@@ -1,0 +1,8 @@
+﻿namespace TheJazMaster.MoreDifficulties;
+
+public interface IMoreDifficultiesApi
+{
+	bool HasAltStarters(Deck deck);
+	StarterDeck? GetAltStarters(Deck deck);
+	bool AreAltStartersEnabled(State state, Deck deck);
+}